# apidog.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EvenIt Backend API",
    "version": "1.0.0",
    "description": "Role-based event marketplace backend for auth, onboarding, listings, bookings, subscriptions, uploads, and admin moderation."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OnboardingMultipartRequest": {
        "type": "object",
        "required": [
          "payload"
        ],
        "properties": {
          "payload": {
            "type": "string",
            "description": "JSON string matching the onboarding request schema for the selected role.",
            "example": "{\"_id\":\"65f1a9d0f1b2c3d4e5f60001\",\"name\":\"Service Provider Example\",\"email\":\"service.provider@example.com\",\"profileInfo\":{\"nidOrTradeLicenseNumber\":\"1234567890123\",\"serviceName\":\"Premium Catering\",\"serviceCategory\":\"Catering\",\"coverageArea\":[\"Dhaka\"],\"verification\":{\"businessType\":\"individual\"}},\"services\":[]}"
          },
          "nationalIdOrTradeLicenseFiles": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
      "EventPlannerProfileInfoUpdateInput": {
        "type": "object",
        "properties": {
          "nidOrTradeLicenseNumber": {
            "type": "string",
            "example": 1234567890123
          },
          "name": {
            "type": "string",
            "example": "Star Events"
          },
          "description": {
            "type": "string",
            "example": "Updated planner description"
          },
          "coverageArea": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Dhaka",
              "Chattogram"
            ]
          },
          "address": {
            "type": "string",
            "example": "Banani, Dhaka"
          },
          "verification": {
            "type": "object",
            "properties": {
              "businessType": {
                "type": "string",
                "enum": [
                  "individual",
                  "company"
                ],
                "example": "company"
              },
              "companyName": {
                "type": "string",
                "example": "Star Events Ltd"
              },
              "nationalIdOrTradeLicenseFiles": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                },
                "example": [
                  "https://example.com/trade-license.pdf"
                ]
              }
            }
          }
        }
      },
...
      "EventPlannerProfileUpdateRequest": {
        "type": "object",
        "description": "Partial update payload for authenticated event planners. Common account fields can be updated at the top level. Planner-specific editable fields live inside `eventPlanner.profileInfo`. There is no common `phone` field on the user model.\n",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Updates both the top-level user fullName and onboarding.eventProvider.fullName.",
            "example": "Star Events"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Updates both the top-level user email and onboarding.eventProvider.email.",
            "example": "planner@example.com"
          },
          "eventPlanner": {
            "type": "object",
            "description": "Role-specific profile fields for event planners.",
            "properties": {
              "profileInfo": {
                "$ref": "#/components/schemas/EventPlannerProfileInfoUpdateInput"
              }
            }
          }
        }
      },
...
      "AuthOnboardingResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Onboarding information submitted successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "onboarding": {
                "type": "object",
                "nullable": true,
                "example": {
                  "submittedAt": "2026-03-24T10:30:00.000Z"
                }
              },
              "user": {
                "$ref": "#/components/schemas/AuthUser"
              }
            }
          }
        }
      },
...
      "BookingListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 10
              },
              "total": {
                "type": "integer",
                "example": 1
              },
              "totalPages": {
                "type": "integer",
                "example": 1
              },
              "hasNextPage": {
                "type": "boolean",
                "example": false
              },
              "hasPrevPage": {
                "type": "boolean",
                "example": false
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            }
          }
        }
      },
...
      "EventPlannerEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65f1a9d0f1b2c3d4e5f60077"
          },
          "fullName": {
            "type": "string",
            "example": "Premium Wedding & Event Planner"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "planner@example.com"
          },
          "role": {
            "type": "string",
            "example": "event_planner"
          },
          "serviceCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "onboarding": {
            "nullable": true,
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventPlannerListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPlannerEntity"
            }
          }
        }
      },
      "PublicErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Service not found"
          }
        }
      },
...
      "PublicEventPlannerListResponse": {
        "$ref": "#/components/schemas/EventPlannerListResponse"
      },
      "PublicEventPlannerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/EventPlannerEntity"
          }
        },
        "example": {
          "success": true,
          "data": {
            "_id": "65f1a9d0f1b2c3d4e5f60077",
            "fullName": "Premium Wedding & Event Planner",
            "email": "planner@example.com",
            "role": "event_planner",
            "serviceCategories": [
              "wedding",
              "corporate"
            ],
            "isEmailVerified": true,
            "onboarding": {
              "verification": {
                "businessType": "company",
                "companyName": "Events Ltd",
                "nationalIdOrTradeLicenseUrl": "https://cdn.example.com/trade-license.pdf"
              },
              "eventProvider": {
                "_id": "65f1a9d0f1b2c3d4e5f60077",
                "fullName": "Premium Wedding & Event Planner",
                "email": "planner@example.com",
                "profileInfo": {
                  "name": "Premium Wedding & Event Planner",
                  "description": "Wedding and corporate event planning",
                  "coverageArea": [
                    "Dhaka",
                    "Chattogram"
                  ],
                  "address": "Banani, Dhaka",
                  "verification": {
                    "businessType": "company",
                    "companyName": "Events Ltd",
                    "nationalIdOrTradeLicenseFiles": [
                      "https://cdn.example.com/trade-license.pdf"
                    ]
                  }
                }
              },
              "submittedAt": "2026-03-20T10:00:00.000Z"
            },
            "createdAt": "2026-03-19T08:00:00.000Z",
            "updatedAt": "2026-03-30T22:41:09.978Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/admin/event-planners": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all event planners",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated event planner list"
          }
        }
      }
    },
...
    "/api/v1/admin/event-planners/{eventPlannerId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get one event planner by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "eventPlannerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event planner details"
          },
          "404": {
            "description": "Event planner not found"
          }
        }
      }
    },
...
    "/api/v1/auth/onboarding/event-planner": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Submit event planner onboarding details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingMultipartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOnboardingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/auth/profile/event-planner": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Update the authenticated event planner profile",
        "description": "Accepts either JSON or multipart form-data. For multipart requests, send the JSON payload in a `payload` field and optional files in `profileImage` and `coverImage`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPlannerProfileUpdateRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "example": "{\"fullName\":\"Star Events\",\"eventPlanner\":{\"profileInfo\":{\"name\":\"Star Events\",\"address\":\"Banani, Dhaka\",\"coverageArea\":[\"Dhaka\",\"Chattogram\"]}}}"
                  },
                  "profileImage": {
                    "type": "string",
                    "format": "binary"
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "400": {
            "description": "Validation error or role mismatch"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Duplicate email"
          }
        }
      }
    },
...
    "/api/v1/bookings/event-planners/{eventPlannerId}": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create a booking request for an event planner",
        "description": "Creates a customer booking request for a verified event planner.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "eventPlannerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bookingDate",
                  "timeSlots"
                ],
                "properties": {
                  "bookingDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-03-20"
                  },
                  "timeSlots": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "14:00"
                    }
                  },
                  "durationHours": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 24
                  },
                  "location": {
                    "type": "string"
                  },
                  "specialInstructions": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event planner booking created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only subscribed customers can create bookings"
          },
          "404": {
            "description": "Event planner not found"
          },
          "409": {
            "description": "Time slot conflict"
          }
        }
      }
    },
...
    "/api/v1/bookings/event-planner": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get event planner booking requests",
        "description": "Returns paginated bookings assigned to the authenticated event planner.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "completed",
                "confirmed",
                "cancelled"
              ]
            },
            "description": "Filter bookings by status. `approved` also matches confirmed bookings."
          }
        ],
        "responses": {
          "200": {
            "description": "Event planner bookings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/event-planners/me/availability": {
      "get": {
        "tags": [
          "EventPlanner"
        ],
        "summary": "Get the authenticated event planner availability calendar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$",
              "example": "2026-04"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event planner availability returned"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only event planners can access this endpoint"
          }
        }
      },
      "patch": {
        "tags": [
          "EventPlanner"
        ],
        "summary": "Block event planner availability hours",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "hours"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-04-18"
                  },
                  "hours": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      10,
                      11,
                      12
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability blocked successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only event planners can access this endpoint"
          }
        }
      },
      "delete": {
        "tags": [
          "EventPlanner"
        ],
        "summary": "Unblock event planner availability hours",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "hours"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-04-18"
                  },
                  "hours": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      10,
                      11
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability unblocked successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only event planners can access this endpoint"
          }
        }
      }
    },
    "/api/v1/event-planners": {
      "get": {
        "tags": [
          "EventPlanner"
        ],
        "summary": "Get all event planners (public)",
        "description": "Returns verified users whose role is `event_planner`.",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated event planner list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPlannerListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/event-planners/{eventPlannerId}": {
      "get": {
        "tags": [
          "EventPlanner"
        ],
        "summary": "Get one event planner by id",
        "description": "Returns one verified public event planner profile.",
        "parameters": [
          {
            "in": "path",
            "name": "eventPlannerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event planner details"
          },
          "404": {
            "description": "Event planner not found"
          }
        }
      }
    },
...
    "/api/v1/public/event-planners": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Get public event planners without authentication",
        "description": "Public read-only endpoint. No request body is accepted. Returns verified event planners who completed event planner onboarding.",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated event planners",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEventPlannerListResponse"
                },
                "example": {
                  "success": true,
                  "meta": {
                    "page": 1,
                    "limit": 10,
                    "total": 1,
                    "totalPages": 1,
                    "hasNextPage": false,
                    "hasPrevPage": false
                  },
                  "data": [
                    {
                      "_id": "65f1a9d0f1b2c3d4e5f60077",
                      "fullName": "Premium Wedding & Event Planner",
                      "email": "planner@example.com",
                      "role": "event_planner",
                      "serviceCategories": [
                        "wedding",
                        "corporate"
                      ],
                      "isEmailVerified": true,
                      "onboarding": {
                        "eventProvider": {
                          "profileInfo": {
                            "name": "Premium Wedding & Event Planner",
                            "coverageArea": [
                              "Dhaka",
                              "Chattogram"
                            ]
                          }
                        }
                      },
                      "createdAt": "2026-03-19T08:00:00.000Z",
                      "updatedAt": "2026-03-30T22:41:09.978Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sorting query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/event-planners/{eventPlannerId}": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Get one event planner without authentication",
        "description": "Public read-only endpoint. No request body is accepted. Returns one event planner by id.",
        "parameters": [
          {
            "in": "path",
            "name": "eventPlannerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event planner details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEventPlannerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid event planner id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event planner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}