# src/modules/event-planner/event-planner.route.ts
import { Router } from 'express';
import { protect } from '../../common/middlewares/auth.middleware';
import { authorize } from '../../common/middlewares/authorize.middleware';
import { validate } from '../../common/middlewares/validate.middleware';
import { EventPlannerController } from './event-planner.controller';
import {
  eventPlannerAvailabilityQuerySchema,
  updateEventPlannerAvailabilitySchema
} from './event-planner.schema';

const router = Router();

/**
 * @openapi
 * /api/v1/event-planners/me/availability:
 *   get:
 *     tags: [EventPlanner]
 *     summary: Get the authenticated event planner availability calendar
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: query
 *         name: month
 *         schema:
 *           type: string
 *           pattern: '^\d{4}-\d{2}$'
 *           example: 2026-04
 *     responses:
 *       200:
 *         description: Event planner availability returned
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Only event planners can access this endpoint
 *   patch:
 *     tags: [EventPlanner]
 *     summary: Block event planner availability hours
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [date, hours]
 *             properties:
 *               date:
 *                 type: string
 *                 format: date
 *                 example: 2026-04-18
 *               hours:
 *                 type: array
 *                 items:
 *                   type: integer
 *                 example: [10, 11, 12]
 *     responses:
 *       200:
 *         description: Availability blocked successfully
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Only event planners can access this endpoint
 *   delete:
 *     tags: [EventPlanner]
 *     summary: Unblock event planner availability hours
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [date, hours]
 *             properties:
 *               date:
 *                 type: string
 *                 format: date
 *                 example: 2026-04-18
 *               hours:
 *                 type: array
 *                 items:
 *                   type: integer
 *                 example: [10, 11]
 *     responses:
 *       200:
 *         description: Availability unblocked successfully
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Only event planners can access this endpoint
 */
router.get(
  '/me/availability',
  protect,
  authorize('event_planner'),
  validate(eventPlannerAvailabilityQuerySchema),
  EventPlannerController.getMyAvailability
);

router.patch(
  '/me/availability',
  protect,
  authorize('event_planner'),
  validate(updateEventPlannerAvailabilitySchema),
  EventPlannerController.blockMyAvailability
);

router.delete(
  '/me/availability',
  protect,
  authorize('event_planner'),
  validate(updateEventPlannerAvailabilitySchema),
  EventPlannerController.unblockMyAvailability
);

/**
 * @openapi
 * tags:
 *   - name: EventPlanner
 *     description: Public listing of verified event providers
 * components:
 *   schemas:
 *     EventPlannerEntity:
 *       type: object
 *       properties:
 *         _id:
 *           type: string
 *           example: 65f1a9d0f1b2c3d4e5f60077
 *         fullName:
 *           type: string
 *           example: Premium Wedding & Event Planner
 *         email:
 *           type: string
 *           format: email
 *           example: planner@example.com
 *         role:
 *           type: string
 *           example: event_planner
 *         serviceCategories:
 *           type: array
 *           items:
 *             type: string
 *         isEmailVerified:
 *           type: boolean
 *           example: true
 *         onboarding:
 *           nullable: true
 *           type: object
 *         createdAt:
 *           type: string
 *           format: date-time
 *         updatedAt:
 *           type: string
 *           format: date-time
 *     EventPlannerListResponse:
 *       type: object
 *       properties:
 *         success:
 *           type: boolean
 *           example: true
 *         meta:
 *           type: object
 *           properties:
 *             page:
 *               type: integer
 *             limit:
 *               type: integer
 *             total:
 *               type: integer
 *             totalPages:
 *               type: integer
 *             hasNextPage:
 *               type: boolean
 *             hasPrevPage:
 *               type: boolean
 *         data:
 *           type: array
 *           items:
 *             $ref: '#/components/schemas/EventPlannerEntity'
 */

/**
 * @openapi
 * /api/v1/event-planners:
 *   get:
 *     tags: [EventPlanner]
 *     summary: Get all event planners (public)
 *     description: Returns verified users whose role is `event_planner`.
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *       - in: query
 *         name: sortBy
 *         schema:
 *           type: string
 *       - in: query
 *         name: sortOrder
 *         schema:
 *           type: string
 *           enum: [asc, desc]
 *     responses:
 *       200:
 *         description: Paginated event planner list
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/EventPlannerListResponse'
 */
router.get('/', EventPlannerController.getAll);

/**
 * @openapi
 * /api/v1/event-planners/{eventPlannerId}:
 *   get:
 *     tags: [EventPlanner]
 *     summary: Get one event planner by id
 *     description: Returns one verified public event planner profile.
 *     parameters:
 *       - in: path
 *         name: eventPlannerId
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Event planner details
 *       404:
 *         description: Event planner not found
 */
router.get('/:eventPlannerId', EventPlannerController.getById);

export const eventPlannerRouter = router;